split up mainactvity, flowfieldactivity and previewactivity
.MainActivity, .flowfield.FlowFieldActivity, and .PreviewActivity are
all seperate now. Lots cleaner :)
